Move binance migration to behind migrations gate
