+++ tests/test_binance_mig.py
@@ -5,7 +5,7 @@ import shutil
 import pytest
 
 from freqtrade.persistence import Trade
-from freqtrade.util.binance_mig import migrate_binance_futures_data, migrate_binance_futures_names
+from freqtrade.util.migrations import migrate_binance_futures_data, migrate_binance_futures_names
 from tests.conftest import create_mock_trades_usdt, log_has
